web-api/requirements.txt
boto3==1.17.49
bcrypt==3.2.0
pyjwt==1.7.1
pynamodb==5.0.3
pynamodb-attributes==0.3.0
dataclasses-json==0.5.2
email-validator==1.1.2
ksuid==1.3
aws-lambda-powertools==1.14.0
packaging==20.8
requests==2.25.1
jsonpickle==2.0.0